+++ ios/Classes/AnotherBrotherPlugin.m
@@ -23,6 +23,7 @@
 #import "Method/EndCommunicationMethodCall.h"
 #import "Method/GetPdfFilePagesMethodCall.h"
 #import "Method/UpdatePrinterSettingsMethodCall.h"
+#import "Method/GetPrinterSettingsMethodCall.h"
 #import "Method/TypeB/TbStartCommunicationMethodCall.h"
 #import "Method/TypeB/TbEndCommunicationMethodCall.h"
 #import "Method/TypeB/TbSendCommandMethodCall.h"
@@ -103,6 +104,9 @@
   }
   else if ([[UpdatePrinterSettingsMethodCall METHOD_NAME] isEqualToString:call.method]) {
         [[[UpdatePrinterSettingsMethodCall alloc] initWithCall:call result:result] execute];
+  }
+  else if ([[GetPrinterSettingsMethodCall METHOD_NAME] isEqualToString:call.method]) {
+        [[[GetPrinterSettingsMethodCall alloc] initWithCall:call result:result] execute];
   }
   // TYPE B
   else if ([[TbStartCommunicationMethodCall METHOD_NAME] isEqualToString:call.method]) {